requirements.txt
asgiref==3.8.1
boto3==1.34.150
botocore==1.34.150
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
django-allauth==0.63.6
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
jmespath==1.0.1
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
<<<<<<< HEAD
s3transfer==0.10.2
six==1.16.0
=======
<<<<<<< HEAD
pillow==10.4.0
=======
>>>>>>> qna
>>>>>>> main
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
